fixed re-export files
