+++ helpers/HttpError.js
@@ -4,4 +4,4 @@ const HttpError = (status, message) => {
   return error;
 };
 
-module.exports = {HttpError};
+module.exports = HttpError;
