+++ middlewares/index.js
@@ -1,0 +1,11 @@
+const isValidId = require("./isValidId");
+const validatePostBody = require("./validatePostBody");
+const validatePutBody = require("./validatePutBody");
+const validateFavorite = require("./validateFavorite");
+
+module.exports = {
+  isValidId,
+  validatePostBody,
+  validatePutBody,
+  validateFavorite,
+};
