+++ middlewares/isValidId.js
@@ -1,10 +1,10 @@
 const { isValidObjectId } = require("mongoose");
-const { httpError } = require("../helpers/HttpError");
+const { HttpError } = require("../helpers");
 
 const isValidId = (req, res, next) => {
   const { contactId } = req.params;
   if (!isValidObjectId(contactId)) {
-    next(httpError(400, `${contactId} is not valid id`));
+    next(HttpError(400, `${contactId} is not valid id`));
   }
   next();
 };
