--- middlewares/validateBody.js
@@ -1,40 +1,0 @@
-const { HttpError } = require("../helpers/HttpError");
-
-const validatePostBody = (schema) => {
-  const func = (req, res, next) => {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      next(HttpError(400, "missing required name field"));
-    }
-    next();
-  };
-  return func;
-};
-
-const validatePutBody = (schema) => {
-  const func = (req, res, next) => {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      next(HttpError(400, "missing field"));
-    }
-    next();
-  };
-  return func;
-};
-
-const validateFavorite = (schema) => {
-  const func = (req, res, next) => {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      next(HttpError(400, "missing field favorite"));
-    }
-    next();
-  };
-  return func;
-};
-
-module.exports = {
-  validatePutBody,
-  validatePostBody,
-  validateFavorite,
-};
