+++ models/contacts.js
@@ -1,7 +1,7 @@
 const { Schema, model } = require("mongoose");
 const Joi = require("joi");
 
-const handleMongooseError = require("../helpers/handleMongooseError");
+const {handleMongooseError} = require("../helpers");
 
 const contactSchema = new Schema(
   {
